refactor: extract achievement and character submodules for maintainability
Achievement module:
- Extract impl Achievements methods into 4 focused submodules:
  unlock.rs, notifications.rs, modal.rs, stats.rs
- types.rs retains struct definitions and tests

Character module:
- Extract input.rs into 4 submodules: creation.rs, select.rs,
  delete.rs, rename.rs
- Extract derived_stats.rs into calculation.rs + multipliers.rs
- Extract prestige.rs into combat_bonuses.rs, tiers.rs,
  prestige_actions.rs

All 1336 tests pass. No behavioral changes — only code moves
with pub use re-exports for backward compatibility.
