+++ src/achievements/types.rs
@@ -301,169 +301,6 @@ pub struct Achievements {
     pub accumulation_start: Option<std::time::Instant>,
 }
 
-impl Achievements {
-    /// Check if an achievement is unlocked.
-    pub fn is_unlocked(&self, id: AchievementId) -> bool {
-        self.unlocked.contains_key(&id)
-    }
-
-    /// Unlock an achievement. Returns true if newly unlocked.
-    pub fn unlock(&mut self, id: AchievementId, character_name: Option<String>) -> bool {
-        if self.is_unlocked(id) {
-            return false;
-        }
-        self.unlocked.insert(
-            id,
-            UnlockedAchievement {
-                unlocked_at: chrono::Utc::now().timestamp(),
-                character_name,
-            },
-        );
-        self.pending_notifications.push(id);
-        self.newly_unlocked.push(id);
-
-        // Add to modal queue and start accumulation timer if not already started
-        self.modal_queue.push(id);
-        if self.accumulation_start.is_none() {
-            self.accumulation_start = Some(std::time::Instant::now());
-        }
-
-        true
-    }
-
-    /// Get the count of pending achievement notifications.
-    pub fn pending_count(&self) -> usize {
-        self.pending_notifications.len()
-    }
-
-    /// Clear pending notifications (call when user views achievements).
-    pub fn clear_pending_notifications(&mut self) {
-        self.recently_unlocked
-            .append(&mut self.pending_notifications);
-    }
-
-    /// Clear recently unlocked list (call when achievement browser closes).
-    pub fn clear_recently_unlocked(&mut self) {
-        self.recently_unlocked.clear();
-    }
-
-    /// Check if an achievement was recently unlocked (for NEW badge in browser).
-    #[allow(dead_code)] // Will be used by achievement browser UI
-    pub fn is_recently_unlocked(&self, id: AchievementId) -> bool {
-        self.recently_unlocked.contains(&id)
-    }
-
-    /// Count recently unlocked achievements in a category (for tab badges).
-    #[allow(dead_code)] // Will be used by achievement browser UI
-    pub fn count_recently_unlocked_by_category(&self, category: AchievementCategory) -> usize {
-        use super::data::ALL_ACHIEVEMENTS;
-        self.recently_unlocked
-            .iter()
-            .filter(|id| {
-                ALL_ACHIEVEMENTS
-                    .iter()
-                    .any(|a| a.id == **id && a.category == category)
-            })
-            .count()
-    }
-
-    /// Take newly unlocked achievements for logging (clears the list).
-    pub fn take_newly_unlocked(&mut self) -> Vec<AchievementId> {
-        std::mem::take(&mut self.newly_unlocked)
-    }
-
-    /// Check if the achievement modal is ready to show.
-    /// Returns true if there are queued achievements and 500ms has elapsed.
-    pub fn is_modal_ready(&self) -> bool {
-        if self.modal_queue.is_empty() {
-            return false;
-        }
-        if let Some(start) = self.accumulation_start {
-            start.elapsed() >= std::time::Duration::from_millis(500)
-        } else {
-            false
-        }
-    }
-
-    /// Take the modal queue for display (clears queue and resets timer).
-    pub fn take_modal_queue(&mut self) -> Vec<AchievementId> {
-        self.accumulation_start = None;
-        std::mem::take(&mut self.modal_queue)
-    }
-
-    /// Update progress on a tracked achievement.
-    pub fn update_progress(&mut self, id: AchievementId, current: u64, target: u64) {
-        self.progress
-            .insert(id, AchievementProgress { current, target });
-    }
-
-    /// Get the progress for an achievement, if any.
-    pub fn get_progress(&self, id: AchievementId) -> Option<&AchievementProgress> {
-        self.progress.get(&id)
-    }
-
-    /// Get the total number of achievements.
-    pub fn total_count(&self) -> usize {
-        use super::data::ALL_ACHIEVEMENTS;
-        ALL_ACHIEVEMENTS.len()
-    }
-
-    /// Get the number of unlocked achievements.
-    pub fn unlocked_count(&self) -> usize {
-        self.unlocked.len()
-    }
-
-    /// Get unlock percentage (0.0 - 100.0).
-    pub fn unlock_percentage(&self) -> f32 {
-        let total = self.total_count();
-        if total == 0 {
-            return 0.0;
-        }
-        (self.unlocked_count() as f32 / total as f32) * 100.0
-    }
-
-    /// Convenience wrapper: unlock with an `Option<&str>` name (avoids repeated `.map(|s| s.to_string())`).
-    pub(super) fn unlock_with_name(
-        &mut self,
-        id: AchievementId,
-        character_name: Option<&str>,
-    ) -> bool {
-        self.unlock(id, character_name.map(|s| s.to_string()))
-    }
-
-    /// Helper to check and unlock milestones. Checks all milestones in order.
-    /// Short-circuits already-unlocked achievements to avoid String allocations
-    /// and HashMap updates on every call.
-    pub(super) fn check_milestones(
-        &mut self,
-        current: u64,
-        milestones: &[(u64, AchievementId)],
-        character_name: Option<&str>,
-    ) {
-        for &(threshold, achievement_id) in milestones {
-            if self.is_unlocked(achievement_id) {
-                continue;
-            }
-            if current >= threshold {
-                self.unlock_with_name(achievement_id, character_name);
-            } else {
-                self.update_progress(achievement_id, current, threshold);
-            }
-        }
-    }
-
-    /// Get count of unlocked/total by category.
-    pub fn count_by_category(&self, category: AchievementCategory) -> (usize, usize) {
-        use super::data::ALL_ACHIEVEMENTS;
-        ALL_ACHIEVEMENTS
-            .iter()
-            .filter(|a| a.category == category)
-            .fold((0, 0), |(unlocked, total), a| {
-                (unlocked + self.is_unlocked(a.id) as usize, total + 1)
-            })
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
