+++ src/character/input.rs
@@ -1,315 +1,21 @@
 //! UI-agnostic input handling for character management screens.
+//!
+//! Each screen has its own sibling module with input enum, result enum, and process function.
+//! This module re-exports all types for backward compatibility.
 
-use crate::ui::character_creation::CharacterCreationScreen;
-use crate::ui::character_delete::CharacterDeleteScreen;
-use crate::ui::character_rename::CharacterRenameScreen;
-use crate::ui::character_select::CharacterSelectScreen;
-
-use super::manager::{CharacterInfo, CharacterManager};
-
-/// Input events for character creation screen.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum CreationInput {
-    /// Character typed
-    Char(char),
-    /// Backspace pressed
-    Backspace,
-    /// Enter pressed to create character
-    Submit,
-    /// Escape pressed to cancel
-    Cancel,
-    /// Any other key
-    Other,
-}
-
-/// Input events for character select screen.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum SelectInput {
-    /// Move selection up
-    Up,
-    /// Move selection down
-    Down,
-    /// Load selected character
-    Select,
-    /// Create new character
-    New,
-    /// Delete selected character
-    Delete,
-    /// Rename selected character
-    Rename,
-    /// Quit the game
-    Quit,
-    /// Any other key
-    Other,
-}
-
-/// Input events for character delete confirmation screen.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum DeleteInput {
-    /// Character typed
-    Char(char),
-    /// Backspace pressed
-    Backspace,
-    /// Enter pressed to confirm deletion
-    Submit,
-    /// Escape pressed to cancel
-    Cancel,
-    /// Any other key
-    Other,
-}
-
-/// Input events for character rename screen.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum RenameInput {
-    /// Character typed
-    Char(char),
-    /// Backspace pressed
-    Backspace,
-    /// Enter pressed to confirm rename
-    Submit,
-    /// Escape pressed to cancel
-    Cancel,
-    /// Any other key
-    Other,
-}
-
-/// Result of processing character creation input.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum CreationResult {
-    /// Stay on creation screen
-    Continue,
-    /// Character created successfully, go to select screen
-    Created,
-    /// Cancelled, go back to select screen (only if characters exist)
-    Cancelled,
-    /// Save failed with error message
-    SaveFailed(String),
-}
-
-/// Result of processing character select input.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum SelectResult {
-    /// Stay on select screen
-    Continue,
-    /// No characters exist, should go to creation
-    NoCharacters,
-    /// Load selected character (returns filename)
-    LoadCharacter(String),
-    /// Go to character creation
-    GoToCreation,
-    /// Go to character delete screen
-    GoToDelete,
-    /// Go to character rename screen
-    GoToRename,
-    /// Quit the game
-    Quit,
-    /// Load failed with error
-    #[allow(dead_code)]
-    LoadFailed(String),
-}
-
-/// Result of processing character delete input.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum DeleteResult {
-    /// Stay on delete screen
-    Continue,
-    /// Character deleted successfully
-    Deleted,
-    /// Cancelled, go back to select screen
-    Cancelled,
-    /// Delete failed with error
-    DeleteFailed(String),
-}
-
-/// Result of processing character rename input.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum RenameResult {
-    /// Stay on rename screen
-    Continue,
-    /// Character renamed successfully
-    Renamed,
-    /// Cancelled, go back to select screen
-    Cancelled,
-    /// Rename failed with error message (sets validation_error)
-    RenameFailed(String),
-}
-
-/// Process input for the character creation screen.
-///
-/// Returns the result of the input processing.
-pub fn process_creation_input(
-    screen: &mut CharacterCreationScreen,
-    input: CreationInput,
-    manager: &CharacterManager,
-    has_existing_characters: bool,
-) -> CreationResult {
-    match input {
-        CreationInput::Char(c) => {
-            screen.handle_char_input(c);
-            CreationResult::Continue
-        }
-        CreationInput::Backspace => {
-            screen.handle_backspace();
-            CreationResult::Continue
-        }
-        CreationInput::Submit => {
-            if screen.is_valid() {
-                let new_name = screen.get_name();
-                let new_state = crate::core::game_state::GameState::new(
-                    new_name,
-                    chrono::Utc::now().timestamp(),
-                );
-                match manager.save_character(&new_state) {
-                    Ok(()) => CreationResult::Created,
-                    Err(e) => {
-                        screen.validation_error = Some(format!("Save failed: {}", e));
-                        CreationResult::SaveFailed(format!("Save failed: {}", e))
-                    }
-                }
-            } else {
-                CreationResult::Continue
-            }
-        }
-        CreationInput::Cancel => {
-            if has_existing_characters {
-                CreationResult::Cancelled
-            } else {
-                CreationResult::Continue
-            }
-        }
-        CreationInput::Other => CreationResult::Continue,
-    }
-}
-
-/// Process input for the character select screen.
-///
-/// Returns the result of the input processing.
-pub fn process_select_input(
-    screen: &mut CharacterSelectScreen,
-    input: SelectInput,
-    characters: &[CharacterInfo],
-) -> SelectResult {
-    if characters.is_empty() {
-        return SelectResult::NoCharacters;
-    }
-
-    // Clamp selected index if needed
-    if screen.selected_index >= characters.len() {
-        screen.selected_index = characters.len().saturating_sub(1);
-    }
-
-    match input {
-        SelectInput::Up => {
-            screen.move_up(characters);
-            SelectResult::Continue
-        }
-        SelectInput::Down => {
-            screen.move_down(characters);
-            SelectResult::Continue
-        }
-        SelectInput::Select => {
-            let selected = &characters[screen.selected_index];
-            if selected.is_corrupted {
-                SelectResult::Continue
-            } else {
-                SelectResult::LoadCharacter(selected.filename.clone())
-            }
-        }
-        SelectInput::New => SelectResult::GoToCreation,
-        SelectInput::Delete => {
-            let selected = &characters[screen.selected_index];
-            if selected.is_corrupted {
-                SelectResult::Continue
-            } else {
-                SelectResult::GoToDelete
-            }
-        }
-        SelectInput::Rename => {
-            let selected = &characters[screen.selected_index];
-            if selected.is_corrupted {
-                SelectResult::Continue
-            } else {
-                SelectResult::GoToRename
-            }
-        }
-        SelectInput::Quit => SelectResult::Quit,
-        SelectInput::Other => SelectResult::Continue,
-    }
-}
-
-/// Process input for the character delete confirmation screen.
-///
-/// Returns the result of the input processing.
-pub fn process_delete_input(
-    screen: &mut CharacterDeleteScreen,
-    input: DeleteInput,
-    manager: &CharacterManager,
-    character: &CharacterInfo,
-) -> DeleteResult {
-    match input {
-        DeleteInput::Char(c) => {
-            screen.handle_char_input(c);
-            DeleteResult::Continue
-        }
-        DeleteInput::Backspace => {
-            screen.handle_backspace();
-            DeleteResult::Continue
-        }
-        DeleteInput::Submit => {
-            if screen.is_confirmed(&character.character_name) {
-                match manager.delete_character(&character.filename) {
-                    Ok(()) => DeleteResult::Deleted,
-                    Err(e) => DeleteResult::DeleteFailed(format!("Failed to delete: {}", e)),
-                }
-            } else {
-                DeleteResult::Continue
-            }
-        }
-        DeleteInput::Cancel => DeleteResult::Cancelled,
-        DeleteInput::Other => DeleteResult::Continue,
-    }
-}
-
-/// Process input for the character rename screen.
-///
-/// Returns the result of the input processing.
-pub fn process_rename_input(
-    screen: &mut CharacterRenameScreen,
-    input: RenameInput,
-    manager: &CharacterManager,
-    character: &CharacterInfo,
-) -> RenameResult {
-    match input {
-        RenameInput::Char(c) => {
-            screen.handle_char_input(c);
-            RenameResult::Continue
-        }
-        RenameInput::Backspace => {
-            screen.handle_backspace();
-            RenameResult::Continue
-        }
-        RenameInput::Submit => {
-            if screen.is_valid() {
-                let new_name = screen.get_name();
-                match manager.rename_character(&character.filename, new_name) {
-                    Ok(()) => RenameResult::Renamed,
-                    Err(e) => {
-                        screen.validation_error = Some(format!("Rename failed: {}", e));
-                        RenameResult::RenameFailed(format!("Rename failed: {}", e))
-                    }
-                }
-            } else {
-                RenameResult::Continue
-            }
-        }
-        RenameInput::Cancel => RenameResult::Cancelled,
-        RenameInput::Other => RenameResult::Continue,
-    }
-}
+pub use super::creation::{process_creation_input, CreationInput, CreationResult};
+pub use super::delete::{process_delete_input, DeleteInput, DeleteResult};
+pub use super::rename::{process_rename_input, RenameInput, RenameResult};
+pub use super::select::{process_select_input, SelectInput, SelectResult};
 
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::character::manager::{CharacterInfo, CharacterManager};
+    use crate::ui::character_creation::CharacterCreationScreen;
+    use crate::ui::character_delete::CharacterDeleteScreen;
+    use crate::ui::character_rename::CharacterRenameScreen;
+    use crate::ui::character_select::CharacterSelectScreen;
 
     // =========================================================================
     // CreationInput tests
