+++ src/character/prestige.rs
@@ -1,265 +1,21 @@
-use crate::core::constants::*;
-use crate::core::game_state::GameState;
+//! Prestige system: tier definitions, combat bonuses, and prestige actions.
+//!
+//! Logic is split across sibling submodules:
+//! - `combat_bonuses.rs` — PrestigeCombatBonuses struct and from_rank()
+//! - `tiers.rs` — PrestigeTier, get_prestige_tier(), tier name lookup
+//! - `prestige_actions.rs` — can_prestige(), perform_prestige(), perform_prestige_with_vault()
 
-/// Combat bonuses from prestige rank (independent of attributes/Haven).
-/// These are flat values that bypass enemy HP scaling.
-#[derive(Debug, Clone, Copy, Default)]
-pub struct PrestigeCombatBonuses {
-    /// Flat damage added after Haven % multiplier, before crit
-    pub flat_damage: u32,
-    /// Flat defense added to DEX-based defense
-    pub flat_defense: u32,
-    /// Percentage points added to crit chance
-    pub crit_chance: f64,
-    /// Flat HP added to combat HP (NOT to DerivedStats.max_hp)
-    pub flat_hp: u32,
-}
-
-impl PrestigeCombatBonuses {
-    pub fn from_rank(rank: u32) -> Self {
-        if rank == 0 {
-            return Self::default();
-        }
-        Self {
-            flat_damage: (PRESTIGE_FLAT_DAMAGE_FACTOR
-                * (rank as f64).powf(PRESTIGE_FLAT_DAMAGE_EXPONENT))
-            .floor() as u32,
-            flat_defense: (PRESTIGE_FLAT_DEFENSE_FACTOR
-                * (rank as f64).powf(PRESTIGE_FLAT_DEFENSE_EXPONENT))
-            .floor() as u32,
-            crit_chance: (rank as f64 * PRESTIGE_CRIT_PER_RANK).min(PRESTIGE_CRIT_CAP),
-            flat_hp: (PRESTIGE_FLAT_HP_FACTOR * (rank as f64).powf(PRESTIGE_FLAT_HP_EXPONENT))
-                .floor() as u32,
-        }
-    }
-}
-
-/// Represents a prestige tier with its properties
-#[derive(Debug, Clone)]
-pub struct PrestigeTier {
-    #[allow(dead_code)]
-    pub rank: u32,
-    pub name: &'static str,
-    pub required_level: u32,
-    pub multiplier: f64,
-}
-
-/// Gets the name for a prestige rank
-fn get_prestige_name(rank: u32) -> &'static str {
-    match rank {
-        0 => "None",
-        // Metals (1-4)
-        1 => "Bronze",
-        2 => "Silver",
-        3 => "Gold",
-        4 => "Platinum",
-        // Gems (5-9)
-        5 => "Diamond",
-        6 => "Emerald",
-        7 => "Sapphire",
-        8 => "Ruby",
-        9 => "Obsidian",
-        // Cosmic (10-14)
-        10 => "Celestial",
-        11 => "Astral",
-        12 => "Cosmic",
-        13 => "Stellar",
-        14 => "Galactic",
-        // Divine (15-19)
-        15 => "Transcendent",
-        16 => "Divine",
-        17 => "Exalted",
-        18 => "Mythic",
-        19 => "Legendary",
-        // Eternal (20+)
-        _ => "Eternal",
-    }
-}
-
-/// Gets the prestige tier for a given rank
-///
-/// # Arguments
-/// * `rank` - The prestige rank
-///
-/// # Returns
-/// The PrestigeTier with name, required level, and multiplier
-///
-/// # Multiplier Formula
-/// Uses diminishing returns: `1 + 0.5 * rank^0.7`
-///
-/// This provides:
-/// - Strong early boost (+50% at P1)
-/// - Tapering gains to prevent late-game trivialization
-/// - Cycles get progressively longer, creating the "wall" feeling
-///
-/// See docs/plans/2026-02-03-prestige-multiplier-rebalance.md for details.
-pub fn get_prestige_tier(rank: u32) -> PrestigeTier {
-    // Diminishing returns formula: 1 + BASE_FACTOR * rank^EXPONENT
-    // P1: 1.5x, P5: 2.5x, P10: 3.5x, P20: 5.1x, P30: 6.4x
-    let multiplier = 1.0 + PRESTIGE_MULT_BASE_FACTOR * (rank as f64).powf(PRESTIGE_MULT_EXPONENT);
-
-    let required_level = match rank {
-        0 => 0,
-        1 => 10,
-        2 => 25,
-        3 => 50,
-        4 => 65,
-        5 => 80,
-        6 => 90,
-        7 => 100,
-        8 => 110,
-        9 => 120,
-        10 => 130,
-        11 => 140,
-        12 => 150,
-        13 => 160,
-        14 => 170,
-        15 => 180,
-        16 => 190,
-        17 => 200,
-        18 => 210,
-        19 => PRESTIGE_HIGH_RANK_BASE_LEVEL,
-        // 20+: continues at +PRESTIGE_HIGH_RANK_LEVEL_STEP per rank
-        _ => {
-            PRESTIGE_HIGH_RANK_BASE_LEVEL
-                + (rank - PRESTIGE_HIGH_RANK_THRESHOLD) * PRESTIGE_HIGH_RANK_LEVEL_STEP
-        }
-    };
-
-    PrestigeTier {
-        rank,
-        name: get_prestige_name(rank),
-        required_level,
-        multiplier,
-    }
-}
-
-/// Gets the next prestige tier based on current rank
-///
-/// # Arguments
-/// * `current_rank` - The player's current prestige rank
-///
-/// # Returns
-/// The PrestigeTier for the next rank
-pub fn get_next_prestige_tier(current_rank: u32) -> PrestigeTier {
-    get_prestige_tier(current_rank + 1)
-}
-
-/// Checks if the player can prestige
-///
-/// # Arguments
-/// * `state` - The current game state
-///
-/// # Returns
-/// true if character level meets the required level for next prestige tier
-pub fn can_prestige(state: &GameState) -> bool {
-    let next_tier = get_next_prestige_tier(state.prestige_rank);
-    state.character_level >= next_tier.required_level
-}
-
-/// Performs a prestige, resetting character progress and incrementing prestige rank
-///
-/// # Arguments
-/// * `state` - The game state to modify
-pub fn perform_prestige(state: &mut GameState) {
-    use super::attributes::Attributes;
-    use crate::combat::CombatState;
-    use crate::items::Equipment;
-
-    // Only prestige if eligible
-    if !can_prestige(state) {
-        return;
-    }
-
-    // Reset character to level 1, XP 0
-    state.character_level = 1;
-    state.character_xp = 0;
-
-    // Reset attributes to base 10
-    state.attributes = Attributes::new();
-
-    // Reset equipment (complete wipe)
-    state.equipment = Equipment::new();
-
-    // Reset active dungeon
-    state.active_dungeon = None;
-
-    // Clear active fishing session (transient state)
-    // Note: Fishing rank and progression (total_fish_caught, legendary_catches, etc.)
-    // are intentionally preserved across prestige as a separate progression track
-    state.active_fishing = None;
-
-    // Clear any active minigame session
-    state.active_minigame = None;
-
-    // Reset combat state with base HP for fresh attributes
-    state.combat_state = CombatState::new(BASE_HP as u32);
-
-    // Increment prestige rank and total prestige count
-    state.prestige_rank += 1;
-    state.total_prestige_count += 1;
-
-    // Clear XP rate tracking so ETA recalculates from fresh post-prestige data
-    state.xp_rate_samples.clear();
-    state.xp_this_second = 0;
-
-    // Reset zone progression but keep unlocks based on new prestige rank
-    state
-        .zone_progression
-        .reset_for_prestige(state.prestige_rank);
-}
-
-/// Performs prestige with Vault item preservation.
-/// `preserved_slots` contains the equipment slots to keep (limited by Vault tier externally).
-pub fn perform_prestige_with_vault(
-    state: &mut GameState,
-    preserved_slots: &[crate::items::EquipmentSlot],
-) {
-    use crate::items::EquipmentSlot;
-
-    if !can_prestige(state) {
-        return;
-    }
-
-    // Save items from preserved slots before reset
-    let mut saved_items: Vec<(EquipmentSlot, crate::items::Item)> = Vec::new();
-    for slot in preserved_slots {
-        if let Some(item) = state.equipment.get(*slot) {
-            saved_items.push((*slot, item.clone()));
-        }
-    }
-
-    // Normal prestige reset
-    perform_prestige(state);
-
-    // Restore preserved items
-    for (slot, item) in saved_items {
-        state.equipment.set(slot, Some(item));
-    }
-}
-
-/// Gets the adventurer rank based on average level
-///
-/// # Arguments
-/// * `avg_level` - The average level across all stats
-///
-/// # Returns
-/// A string describing the adventurer's rank
-pub fn get_adventurer_rank(avg_level: u32) -> &'static str {
-    match avg_level {
-        0..=9 => "Novice",
-        10..=24 => "Adept",
-        25..=49 => "Master",
-        50..=74 => "Grand Master",
-        75..=99 => "Legend",
-        _ => "Mythic",
-    }
-}
+pub use super::combat_bonuses::PrestigeCombatBonuses;
+pub use super::prestige_actions::{can_prestige, perform_prestige, perform_prestige_with_vault};
+pub use super::tiers::{
+    get_adventurer_rank, get_next_prestige_tier, get_prestige_tier, PrestigeTier,
+};
 
 #[cfg(test)]
 mod tests {
     use super::*;
     use crate::character::attributes::AttributeType;
+    use crate::core::game_state::GameState;
 
     #[test]
     fn test_get_prestige_tier() {
